tests: Make dedicated Heap Untyped files
