+++ tests/unit/descriptor_heap_positive.cpp
@@ -18,8 +18,6 @@
 #include "../framework/layer_validation_tests.h"
 #include "../framework/pipeline_helper.h"
 
-constexpr uint32_t kMaxSSBO = 128;  // max bufferDescriptorSize is allowed to be
-
 class PositiveDescriptorHeap : public DescriptorHeapTest {};
 
 void DescriptorHeapTest::InitBasicDescriptorHeap() {
@@ -1807,84 +1805,6 @@ TEST_F(PositiveDescriptorHeap, MappingSourceIndirectAddress) {
     }
 }
 
-TEST_F(PositiveDescriptorHeap, UntypedPointerBuffer) {
-    AddRequiredExtensions(VK_KHR_SHADER_UNTYPED_POINTERS_EXTENSION_NAME);
-    AddRequiredFeature(vkt::Feature::shaderUntypedPointers);
-    RETURN_IF_SKIP(InitBasicDescriptorHeap());
-
-    const VkDeviceSize resource_stride = heap_props.bufferDescriptorSize;
-    CreateResourceHeap(resource_stride);
-
-    vkt::Buffer buffer(*m_device, 256, VK_BUFFER_USAGE_2_STORAGE_BUFFER_BIT_KHR, vkt::device_address);
-
-    VkHostAddressRangeEXT descriptor_host = {resource_heap_data_, static_cast<size_t>(resource_stride)};
-    VkDeviceAddressRangeEXT device_range = {buffer.Address(), 256};
-
-    VkResourceDescriptorInfoEXT descriptor_info = vku::InitStructHelper();
-    descriptor_info.type = VK_DESCRIPTOR_TYPE_STORAGE_BUFFER;
-    descriptor_info.data.pAddressRange = &device_range;
-
-    vk::WriteResourceDescriptorsEXT(*m_device, 1u, &descriptor_info, &descriptor_host);
-
-    VkDescriptorSetAndBindingMappingEXT mapping = MakeSetAndBindingMapping(0, 0);
-    mapping.source = VK_DESCRIPTOR_MAPPING_SOURCE_HEAP_WITH_CONSTANT_OFFSET_EXT;
-    mapping.sourceData.constantOffset.heapOffset = 0;
-    mapping.sourceData.constantOffset.heapArrayStride = 0;
-
-    VkShaderDescriptorSetAndBindingMappingInfoEXT mapping_info = vku::InitStructHelper();
-    mapping_info.mappingCount = 1u;
-    mapping_info.pMappings = &mapping;
-
-    const std::string spirv = R"(
-        OpCapability Shader
-        OpCapability UntypedPointersKHR
-        OpExtension "SPV_KHR_untyped_pointers"
-        OpMemoryModel Logical GLSL450
-        OpEntryPoint GLCompute %main "main" %var
-        OpExecutionMode %main LocalSize 1 1 1
-        OpDecorate %block Block
-        OpMemberDecorate %block 0 Offset 0
-        OpDecorate %var DescriptorSet 0
-        OpDecorate %var Binding 0
-        %void = OpTypeVoid
-        %void_fn = OpTypeFunction %void
-        %int = OpTypeInt 32 0
-        %int_0 = OpConstant %int 0
-        %int_1 = OpConstant %int 1
-        %block = OpTypeStruct %int
-        %ptr = OpTypeUntypedPointerKHR StorageBuffer
-        %var = OpUntypedVariableKHR %ptr StorageBuffer %block
-        %main = OpFunction %void None %void_fn
-        %entry = OpLabel
-        %access = OpUntypedAccessChainKHR %ptr %block %var %int_0
-        OpAtomicStore %access %int_1 %int_0 %int_1
-        OpReturn
-        OpFunctionEnd
-    )";
-
-    VkShaderObj cs_module = VkShaderObj::CreateFromASM(this, spirv.c_str(), VK_SHADER_STAGE_COMPUTE_BIT, SPV_ENV_VULKAN_1_2);
-
-    VkPipelineCreateFlags2CreateInfoKHR pipeline_create_flags_2_create_info = vku::InitStructHelper();
-    pipeline_create_flags_2_create_info.flags = VK_PIPELINE_CREATE_2_DESCRIPTOR_HEAP_BIT_EXT;
-
-    CreateComputePipelineHelper pipe(*this, &pipeline_create_flags_2_create_info);
-    pipe.cp_ci_.layout = VK_NULL_HANDLE;
-    pipe.cp_ci_.stage = cs_module.GetStageCreateInfo(&mapping_info);
-    pipe.CreateComputePipeline(false);
-
-    m_command_buffer.Begin();
-    BindResourceHeap();
-    vk::CmdBindPipeline(m_command_buffer, VK_PIPELINE_BIND_POINT_COMPUTE, pipe);
-    vk::CmdDispatch(m_command_buffer, 1, 1, 1);
-    m_command_buffer.End();
-    m_default_queue->SubmitAndWait(m_command_buffer);
-
-    if (!IsPlatformMockICD()) {
-        uint32_t* data = static_cast<uint32_t*>(buffer.Memory().Map());
-        ASSERT_EQ(data[0], 1u);
-    }
-}
-
 TEST_F(PositiveDescriptorHeap, NestedResourceInheritance) {
     TEST_DESCRIPTION("Validate that inherited ranges match primary buffer");
     AddRequiredExtensions(VK_EXT_NESTED_COMMAND_BUFFER_EXTENSION_NAME);
@@ -2494,330 +2414,6 @@ TEST_F(PositiveDescriptorHeap, ResourceMaskSameBinding) {
     m_default_queue->SubmitAndWait(m_command_buffer);
 }
 
-TEST_F(PositiveDescriptorHeap, ArrayStrideIdEXT) {
-    SetTargetApiVersion(VK_API_VERSION_1_4);
-    AddRequiredExtensions(VK_KHR_SHADER_UNTYPED_POINTERS_EXTENSION_NAME);
-    AddRequiredFeature(vkt::Feature::shaderUntypedPointers);
-    RETURN_IF_SKIP(InitBasicDescriptorHeap());
-
-    std::stringstream cs_source;
-    cs_source << R"(
-               OpCapability Shader
-               OpCapability UntypedPointersKHR
-               OpCapability DescriptorHeapEXT
-               OpCapability ImageBuffer
-               OpExtension "SPV_KHR_untyped_pointers"
-               OpExtension "SPV_EXT_descriptor_heap"
-               OpMemoryModel Logical GLSL450
-               OpEntryPoint GLCompute %1 "main" %2
-               OpExecutionMode %1 LocalSize 1 1 1
-               OpDecorate %2 BuiltIn ResourceHeapEXT
-               OpDecorate %3 SpecId 0
-               OpDecorateId %4 ArrayStrideIdEXT %3
-          %5 = OpTypeVoid
-          %6 = OpTypeFunction %5
-          %7 = OpTypeInt 32 0
-          %8 = OpConstant %7 0
-          %9 = OpConstant %7 2
-         %10 = OpConstant %7 51966
-         %11 = OpConstant %7 0
-          %3 = OpSpecConstant %7 0
-         %12 = OpTypeUntypedPointerKHR UniformConstant
-          %2 = OpUntypedVariableKHR %12 UniformConstant
-         %13 = OpTypeImage %7 Buffer 0 0 0 2 R32ui
-          %4 = OpTypeRuntimeArray %13
-          %1 = OpFunction %5 None %6
-         %14 = OpLabel
-         %15 = OpUntypedAccessChainKHR %12 %4 %2 %9
-         %16 = OpLoad %13 %15
-               OpImageWrite %16 %8 %10
-               OpReturn
-               OpFunctionEnd
-    )";
-
-    const uint32_t data = (uint32_t)heap_props.imageDescriptorSize;
-    const VkSpecializationMapEntry entry = {0, 0, sizeof(uint32_t)};
-    const VkSpecializationInfo specialization_info = {1, &entry, sizeof(uint32_t), &data};
-
-    VkPipelineCreateFlags2CreateInfo pipeline_create_flags_2_create_info = vku::InitStructHelper();
-    pipeline_create_flags_2_create_info.flags = VK_PIPELINE_CREATE_2_DESCRIPTOR_HEAP_BIT_EXT;
-
-    VkShaderObj cs_module = VkShaderObj(*m_device, cs_source.str().c_str(), VK_SHADER_STAGE_COMPUTE_BIT, SPV_ENV_VULKAN_1_4,
-                                        SPV_SOURCE_ASM, &specialization_info);
-
-    CreateComputePipelineHelper pipe(*this, &pipeline_create_flags_2_create_info);
-    pipe.cp_ci_.layout = VK_NULL_HANDLE;
-    pipe.cp_ci_.stage = cs_module.GetStageCreateInfo();
-    pipe.CreateComputePipeline(false);
-}
-
-TEST_F(PositiveDescriptorHeap, OffsetIdEXT) {
-    SetTargetApiVersion(VK_API_VERSION_1_4);
-    AddRequiredExtensions(VK_KHR_SHADER_UNTYPED_POINTERS_EXTENSION_NAME);
-    AddRequiredFeature(vkt::Feature::shaderUntypedPointers);
-    RETURN_IF_SKIP(InitBasicDescriptorHeap());
-
-    std::stringstream cs_source;
-    cs_source << R"(
-               OpCapability Shader
-               OpCapability UntypedPointersKHR
-               OpCapability DescriptorHeapEXT
-               OpCapability Sampled1D
-               OpExtension "SPV_KHR_untyped_pointers"
-               OpExtension "SPV_EXT_descriptor_heap"
-               OpMemoryModel Logical GLSL450
-               OpEntryPoint GLCompute %1 "main" %2 %3
-               OpExecutionMode %1 LocalSize 1 1 1
-               OpDecorate %4 SpecId 0
-               OpDecorate %5 SpecId 1
-               OpDecorate %6 SpecId 2
-               OpDecorate %2 BuiltIn ResourceHeapEXT
-               OpDecorate %3 BuiltIn SamplerHeapEXT
-               OpDecorateId %7 OffsetIdEXT %4
-               OpDecorateId %8 OffsetIdEXT %5
-               OpDecorateId %9 OffsetIdEXT %6
-               OpMemberDecorate %10 0 Offset 0
-         %11 = OpTypeVoid
-         %12 = OpTypeFunction %11
-         %13 = OpTypeInt 32 0
-         %14 = OpTypeVector %13 4
-         %15 = OpTypeFloat 32
-         %16 = OpConstant %13 0
-         %17 = OpConstant %13 1
-         %18 = OpConstant %15 0
-         %19 = OpConstant %15 1
-          %4 = OpSpecConstant %13 0
-          %5 = OpSpecConstant %13 0
-          %6 = OpSpecConstant %13 0
-         %20 = OpTypeImage %13 1D 0 0 0 1 Unknown
-         %21 = OpTypeBufferEXT StorageBuffer
-         %22 = OpTypeSampler
-         %23 = OpTypeSampledImage %20
-          %7 = OpTypeRuntimeArray %20
-          %8 = OpTypeRuntimeArray %21
-          %9 = OpTypeRuntimeArray %22
-         %10 = OpTypeStruct %13
-         %24 = OpTypeUntypedPointerKHR UniformConstant
-         %25 = OpTypeUntypedPointerKHR StorageBuffer
-          %2 = OpUntypedVariableKHR %24 UniformConstant
-          %3 = OpUntypedVariableKHR %24 UniformConstant
-          %1 = OpFunction %11 None %12
-         %26 = OpLabel
-         %27 = OpUntypedAccessChainKHR %24 %7 %2 %16
-         %28 = OpLoad %20 %27
-         %29 = OpUntypedAccessChainKHR %24 %9 %3 %16
-         %30 = OpLoad %22 %29
-         %31 = OpSampledImage %23 %28 %30
-         %32 = OpImageSampleExplicitLod %14 %31 %19 Lod %18
-         %33 = OpCompositeExtract %13 %32 0
-         %34 = OpUntypedAccessChainKHR %24 %8 %2 %16
-         %35 = OpBufferPointerEXT %25 %34
-               OpStore %35 %33
-               OpReturn
-               OpFunctionEnd
-    )";
-
-    VkPipelineCreateFlags2CreateInfo pipeline_create_flags_2_create_info = vku::InitStructHelper();
-    pipeline_create_flags_2_create_info.flags = VK_PIPELINE_CREATE_2_DESCRIPTOR_HEAP_BIT_EXT;
-
-    VkShaderObj cs_module =
-        VkShaderObj(*m_device, cs_source.str().c_str(), VK_SHADER_STAGE_COMPUTE_BIT, SPV_ENV_VULKAN_1_4, SPV_SOURCE_ASM);
-
-    VkShaderDescriptorSetAndBindingMappingInfoEXT mapping_info = vku::InitStructHelper();
-
-    CreateComputePipelineHelper pipe(*this, &pipeline_create_flags_2_create_info);
-    pipe.cp_ci_.layout = VK_NULL_HANDLE;
-    pipe.cp_ci_.stage = cs_module.GetStageCreateInfo(&mapping_info);
-    pipe.CreateComputePipeline(false);
-}
-
-TEST_F(PositiveDescriptorHeap, LayoutDescriptorHeap) {
-    AddRequiredExtensions(VK_KHR_SHADER_UNTYPED_POINTERS_EXTENSION_NAME);
-    AddRequiredFeature(vkt::Feature::shaderUntypedPointers);
-    RETURN_IF_SKIP(InitBasicDescriptorHeap());
-
-    const VkDeviceSize resource_stride = heap_props.bufferDescriptorSize;
-    CreateResourceHeap(resource_stride);
-
-    vkt::Buffer buffer(*m_device, 256, VK_BUFFER_USAGE_2_STORAGE_BUFFER_BIT_KHR, vkt::device_address);
-
-    VkHostAddressRangeEXT descriptor_host = {resource_heap_data_, static_cast<size_t>(resource_stride)};
-    VkDeviceAddressRangeEXT device_range = {buffer.Address(), 256};
-    VkResourceDescriptorInfoEXT descriptor_info = vku::InitStructHelper();
-    descriptor_info.type = VK_DESCRIPTOR_TYPE_STORAGE_BUFFER;
-    descriptor_info.data.pAddressRange = &device_range;
-
-    vk::WriteResourceDescriptorsEXT(*m_device, 1u, &descriptor_info, &descriptor_host);
-
-    char const* cs_source = R"glsl(
-        #version 450
-        #extension GL_EXT_descriptor_heap : require
-        layout(local_size_x = 1) in;
-        layout(descriptor_heap) buffer A { uint a; } heap[];
-        layout(push_constant) uniform PushConstant {
-            uint b;
-        };
-        void main() {
-            heap[0].a = b;
-        }
-    )glsl";
-    VkShaderObj cs_module = VkShaderObj(*m_device, cs_source, VK_SHADER_STAGE_COMPUTE_BIT, SPV_ENV_VULKAN_1_2);
-
-    VkPipelineCreateFlags2CreateInfoKHR pipeline_create_flags_2_create_info = vku::InitStructHelper();
-    pipeline_create_flags_2_create_info.flags = VK_PIPELINE_CREATE_2_DESCRIPTOR_HEAP_BIT_EXT;
-
-    CreateComputePipelineHelper pipe(*this, &pipeline_create_flags_2_create_info);
-    pipe.cp_ci_.layout = VK_NULL_HANDLE;
-    pipe.cp_ci_.stage = cs_module.GetStageCreateInfo();
-    pipe.CreateComputePipeline(false);
-
-    uint32_t src_data = 4321u;
-
-    VkPushDataInfoEXT push_data_info = vku::InitStructHelper();
-    push_data_info.offset = 0u;
-    push_data_info.data.size = sizeof(uint32_t);
-    push_data_info.data.address = &src_data;
-
-    VkPushConstantRange push_const_range = {VK_SHADER_STAGE_COMPUTE_BIT, 0, sizeof(uint32_t)};
-    VkPipelineLayoutCreateInfo pipeline_layout_info = vku::InitStructHelper();
-    pipeline_layout_info.pushConstantRangeCount = 1u;
-    pipeline_layout_info.pPushConstantRanges = &push_const_range;
-    vkt::PipelineLayout pipeline_layout(*m_device, pipeline_layout_info);
-
-    m_command_buffer.Begin();
-    vk::CmdPushConstants(m_command_buffer, pipeline_layout.handle(), VK_SHADER_STAGE_COMPUTE_BIT, 0, sizeof(uint32_t), &src_data);
-    BindResourceHeap();
-    vk::CmdPushDataEXT(m_command_buffer, &push_data_info);
-    vk::CmdBindPipeline(m_command_buffer, VK_PIPELINE_BIND_POINT_COMPUTE, pipe);
-    vk::CmdDispatch(m_command_buffer, 1, 1, 1);
-    m_command_buffer.End();
-    m_default_queue->SubmitAndWait(m_command_buffer);
-
-    if (!IsPlatformMockICD()) {
-        uint32_t* data = static_cast<uint32_t*>(buffer.Memory().Map());
-        ASSERT_EQ(data[0], src_data);
-    }
-}
-
-TEST_F(PositiveDescriptorHeap, UntypedImageAndSampler) {
-    AddRequiredExtensions(VK_KHR_SHADER_UNTYPED_POINTERS_EXTENSION_NAME);
-    AddRequiredFeature(vkt::Feature::shaderUntypedPointers);
-    RETURN_IF_SKIP(InitBasicDescriptorHeap());
-    InitRenderTarget();
-
-    const uint32_t buffer_index = 16u;
-
-    const VkDeviceSize image_offset = 0u;
-    const VkDeviceSize image_size = heap_props.imageDescriptorSize;
-    const VkDeviceSize buffer_offset = heap_props.bufferDescriptorSize * buffer_index;
-    const VkDeviceSize buffer_size = heap_props.bufferDescriptorSize;
-    const VkDeviceSize resource_heap_app_size = buffer_offset + buffer_size;
-
-    CreateResourceHeap(resource_heap_app_size);
-
-    vkt::Buffer buffer(*m_device, 256, VK_BUFFER_USAGE_2_STORAGE_BUFFER_BIT_KHR, vkt::device_address);
-    vkt::Image image(*m_device, 32u, 32u, VK_FORMAT_R8G8B8A8_UNORM, VK_IMAGE_USAGE_TRANSFER_DST_BIT | VK_IMAGE_USAGE_SAMPLED_BIT);
-
-    VkHostAddressRangeEXT resource_host[2];
-    resource_host[0].address = resource_heap_data_ + image_offset;
-    resource_host[0].size = static_cast<size_t>(image_size);
-    resource_host[1].address = resource_heap_data_ + buffer_offset;
-    resource_host[1].size = static_cast<size_t>(buffer_size);
-
-    VkImageViewCreateInfo view_info = image.BasicViewCreatInfo(VK_IMAGE_ASPECT_COLOR_BIT);
-
-    VkImageDescriptorInfoEXT image_info = vku::InitStructHelper();
-    image_info.pView = &view_info;
-    image_info.layout = VK_IMAGE_LAYOUT_SHADER_READ_ONLY_OPTIMAL;
-
-    VkDeviceAddressRangeEXT buffer_address_range = {buffer.Address(), 256};
-
-    VkResourceDescriptorInfoEXT descriptor_info[2];
-    descriptor_info[0] = vku::InitStructHelper();
-    descriptor_info[0].type = VK_DESCRIPTOR_TYPE_SAMPLED_IMAGE;
-    descriptor_info[0].data.pImage = &image_info;
-    descriptor_info[1] = vku::InitStructHelper();
-    descriptor_info[1].type = VK_DESCRIPTOR_TYPE_STORAGE_BUFFER;
-    descriptor_info[1].data.pAddressRange = &buffer_address_range;
-    vk::WriteResourceDescriptorsEXT(*m_device, 2u, descriptor_info, resource_host);
-
-    VkDeviceSize sampler_desc_heap_size_tracker = 0u;
-    const VkDeviceSize sampler_offset = AlignedAppend(sampler_desc_heap_size_tracker, VK_DESCRIPTOR_TYPE_SAMPLER);
-    const VkDeviceSize sampler_size = sampler_desc_heap_size_tracker - sampler_offset;
-
-    CreateSamplerHeap(sampler_desc_heap_size_tracker);
-
-    VkSamplerCreateInfo sampler_info = SafeSaneSamplerCreateInfo();
-
-    VkHostAddressRangeEXT sampler_host = {sampler_heap_data_ + sampler_offset, static_cast<size_t>(sampler_size)};
-    vk::WriteSamplerDescriptorsEXT(*m_device, 1u, &sampler_info, &sampler_host);
-
-    char const* cs_source = R"glsl(
-        #version 450
-        #extension GL_EXT_descriptor_heap : require
-        layout(descriptor_heap) uniform texture2D heapTextures[];
-        layout(descriptor_heap) uniform sampler heapSamplers[];
-        layout(descriptor_heap) buffer ssbo {
-            vec4 data;
-        } heapBuffer[];
-        void main() {
-            heapBuffer[16].data = texture(sampler2D(heapTextures[0], heapSamplers[0]), vec2(0.5f));
-        }
-    )glsl";
-    VkShaderObj cs_module = VkShaderObj(*m_device, cs_source, VK_SHADER_STAGE_COMPUTE_BIT, SPV_ENV_VULKAN_1_2);
-
-    VkPipelineCreateFlags2CreateInfoKHR pipeline_create_flags_2_create_info = vku::InitStructHelper();
-    pipeline_create_flags_2_create_info.flags = VK_PIPELINE_CREATE_2_DESCRIPTOR_HEAP_BIT_EXT;
-
-    CreateComputePipelineHelper pipe(*this, &pipeline_create_flags_2_create_info);
-    pipe.cp_ci_.layout = VK_NULL_HANDLE;
-    pipe.cp_ci_.stage = cs_module.GetStageCreateInfo();
-    pipe.CreateComputePipeline(false);
-
-    m_command_buffer.Begin();
-
-    VkImageMemoryBarrier image_barrier = vku::InitStructHelper();
-    image_barrier.srcAccessMask = VK_ACCESS_NONE;
-    image_barrier.dstAccessMask = VK_ACCESS_TRANSFER_WRITE_BIT;
-    image_barrier.oldLayout = VK_IMAGE_LAYOUT_UNDEFINED;
-    image_barrier.newLayout = VK_IMAGE_LAYOUT_TRANSFER_DST_OPTIMAL;
-    image_barrier.srcQueueFamilyIndex = VK_QUEUE_FAMILY_IGNORED;
-    image_barrier.dstQueueFamilyIndex = VK_QUEUE_FAMILY_IGNORED;
-    image_barrier.image = image;
-    image_barrier.subresourceRange = {VK_IMAGE_ASPECT_COLOR_BIT, 0u, 1u, 0u, 1u};
-    vk::CmdPipelineBarrier(m_command_buffer, VK_PIPELINE_STAGE_TOP_OF_PIPE_BIT, VK_PIPELINE_STAGE_TRANSFER_BIT, 0u, 0u, nullptr, 0u,
-                           nullptr, 1u, &image_barrier);
-
-    VkClearColorValue color = {};
-    color.float32[0] = 0.2f;
-    color.float32[1] = 0.4f;
-    color.float32[2] = 0.6f;
-    color.float32[3] = 0.8f;
-    vk::CmdClearColorImage(m_command_buffer, image, VK_IMAGE_LAYOUT_TRANSFER_DST_OPTIMAL, &color, 1u,
-                           &image_barrier.subresourceRange);
-
-    image_barrier.srcAccessMask = image_barrier.dstAccessMask;
-    image_barrier.dstAccessMask = VK_ACCESS_SHADER_READ_BIT;
-    image_barrier.oldLayout = image_barrier.newLayout;
-    image_barrier.newLayout = VK_IMAGE_LAYOUT_SHADER_READ_ONLY_OPTIMAL;
-    vk::CmdPipelineBarrier(m_command_buffer, VK_PIPELINE_STAGE_TRANSFER_BIT, VK_PIPELINE_STAGE_COMPUTE_SHADER_BIT, 0u, 0u, nullptr,
-                           0u, nullptr, 1u, &image_barrier);
-
-    vk::CmdBindPipeline(m_command_buffer, VK_PIPELINE_BIND_POINT_COMPUTE, pipe);
-    BindResourceHeap();
-    BindSamplerHeap();
-    vk::CmdDispatch(m_command_buffer, 1u, 1u, 1u);
-    m_command_buffer.End();
-    m_default_queue->SubmitAndWait(m_command_buffer);
-
-    if (!IsPlatformMockICD()) {
-        float* data = static_cast<float*>(buffer.Memory().Map());
-        for (uint32_t i = 0; i < 4; ++i) {
-            ASSERT_EQ(data[i], color.float32[i]);
-        }
-    }
-}
-
 TEST_F(PositiveDescriptorHeap, UboAndSsboBindings) {
     RETURN_IF_SKIP(InitBasicDescriptorHeap());
 
@@ -2966,533 +2562,6 @@ TEST_F(PositiveDescriptorHeap, NonConstantMemoryAccess) {
         pipe.cp_ci_.layout = VK_NULL_HANDLE;
         pipe.cp_ci_.stage = cs_module.GetStageCreateInfo(&mapping_info);
         pipe.CreateComputePipeline(false);
-    }
-}
-
-TEST_F(PositiveDescriptorHeap, UntypedStorageImage) {
-    AddRequiredExtensions(VK_KHR_SHADER_UNTYPED_POINTERS_EXTENSION_NAME);
-    AddRequiredFeature(vkt::Feature::shaderUntypedPointers);
-    RETURN_IF_SKIP(InitBasicDescriptorHeap());
-    InitRenderTarget();
-
-    const uint32_t buffer_index = 16u;
-
-    const VkDeviceSize image_offset = 0u;
-    const VkDeviceSize image_size = heap_props.imageDescriptorSize;
-    const VkDeviceSize buffer_offset = heap_props.bufferDescriptorSize * buffer_index;
-    const VkDeviceSize buffer_size = heap_props.bufferDescriptorSize;
-    const VkDeviceSize resource_heap_app_size = buffer_offset + buffer_size;
-
-    CreateResourceHeap(resource_heap_app_size);
-
-    vkt::Buffer buffer(*m_device, sizeof(float) * 4u, VK_BUFFER_USAGE_2_STORAGE_BUFFER_BIT_KHR, vkt::device_address);
-    vkt::Image image(*m_device, 32u, 32u, VK_FORMAT_R8G8B8A8_SINT, VK_IMAGE_USAGE_TRANSFER_DST_BIT | VK_IMAGE_USAGE_STORAGE_BIT);
-
-    VkHostAddressRangeEXT resource_host[2];
-    resource_host[0].address = resource_heap_data_ + image_offset;
-    resource_host[0].size = static_cast<size_t>(image_size);
-    resource_host[1].address = resource_heap_data_ + buffer_offset;
-    resource_host[1].size = static_cast<size_t>(buffer_size);
-
-    VkImageViewCreateInfo view_info = image.BasicViewCreatInfo(VK_IMAGE_ASPECT_COLOR_BIT);
-
-    VkImageDescriptorInfoEXT image_info = vku::InitStructHelper();
-    image_info.pView = &view_info;
-    image_info.layout = VK_IMAGE_LAYOUT_SHADER_READ_ONLY_OPTIMAL;
-
-    VkDeviceAddressRangeEXT buffer_address_range = buffer.AddressRange();
-
-    VkResourceDescriptorInfoEXT descriptor_info[2];
-    descriptor_info[0] = vku::InitStructHelper();
-    descriptor_info[0].type = VK_DESCRIPTOR_TYPE_STORAGE_IMAGE;
-    descriptor_info[0].data.pImage = &image_info;
-    descriptor_info[1] = vku::InitStructHelper();
-    descriptor_info[1].type = VK_DESCRIPTOR_TYPE_STORAGE_BUFFER;
-    descriptor_info[1].data.pAddressRange = &buffer_address_range;
-    vk::WriteResourceDescriptorsEXT(*m_device, 2u, descriptor_info, resource_host);
-
-    char const* cs_source = R"glsl(
-        #version 450
-        #extension GL_EXT_descriptor_heap : require
-        layout(descriptor_heap, rgba8i) uniform iimage2D heapImages[];
-        layout(descriptor_heap) buffer ssbo {
-	        ivec4 data;
-        } heapBuffer[];
-        void main() {
-	        heapBuffer[16].data = imageLoad(heapImages[0], ivec2(0));
-        }
-    )glsl";
-    VkShaderObj cs_module = VkShaderObj(*m_device, cs_source, VK_SHADER_STAGE_COMPUTE_BIT, SPV_ENV_VULKAN_1_2);
-
-    VkPipelineCreateFlags2CreateInfoKHR pipeline_create_flags_2_create_info = vku::InitStructHelper();
-    pipeline_create_flags_2_create_info.flags = VK_PIPELINE_CREATE_2_DESCRIPTOR_HEAP_BIT_EXT;
-
-    CreateComputePipelineHelper pipe(*this, &pipeline_create_flags_2_create_info);
-    pipe.cp_ci_.layout = VK_NULL_HANDLE;
-    pipe.cp_ci_.stage = cs_module.GetStageCreateInfo();
-    pipe.CreateComputePipeline(false);
-
-    m_command_buffer.Begin();
-
-    VkImageMemoryBarrier image_barrier = vku::InitStructHelper();
-    image_barrier.srcAccessMask = VK_ACCESS_NONE;
-    image_barrier.dstAccessMask = VK_ACCESS_TRANSFER_WRITE_BIT;
-    image_barrier.oldLayout = VK_IMAGE_LAYOUT_UNDEFINED;
-    image_barrier.newLayout = VK_IMAGE_LAYOUT_TRANSFER_DST_OPTIMAL;
-    image_barrier.srcQueueFamilyIndex = VK_QUEUE_FAMILY_IGNORED;
-    image_barrier.dstQueueFamilyIndex = VK_QUEUE_FAMILY_IGNORED;
-    image_barrier.image = image;
-    image_barrier.subresourceRange = {VK_IMAGE_ASPECT_COLOR_BIT, 0u, 1u, 0u, 1u};
-    vk::CmdPipelineBarrier(m_command_buffer, VK_PIPELINE_STAGE_TOP_OF_PIPE_BIT, VK_PIPELINE_STAGE_TRANSFER_BIT, 0u, 0u, nullptr, 0u,
-                           nullptr, 1u, &image_barrier);
-
-    VkClearColorValue color = {};
-    color.int32[0] = 1;
-    color.int32[1] = 2;
-    color.int32[2] = 3;
-    color.int32[3] = 4;
-    vk::CmdClearColorImage(m_command_buffer, image, VK_IMAGE_LAYOUT_TRANSFER_DST_OPTIMAL, &color, 1u,
-                           &image_barrier.subresourceRange);
-
-    image_barrier.srcAccessMask = image_barrier.dstAccessMask;
-    image_barrier.dstAccessMask = VK_ACCESS_SHADER_READ_BIT;
-    image_barrier.oldLayout = image_barrier.newLayout;
-    image_barrier.newLayout = VK_IMAGE_LAYOUT_GENERAL;
-    vk::CmdPipelineBarrier(m_command_buffer, VK_PIPELINE_STAGE_TRANSFER_BIT, VK_PIPELINE_STAGE_COMPUTE_SHADER_BIT, 0u, 0u, nullptr,
-                           0u, nullptr, 1u, &image_barrier);
-
-    vk::CmdBindPipeline(m_command_buffer, VK_PIPELINE_BIND_POINT_COMPUTE, pipe);
-    BindResourceHeap();
-    vk::CmdDispatch(m_command_buffer, 1u, 1u, 1u);
-    m_command_buffer.End();
-    m_default_queue->SubmitAndWait(m_command_buffer);
-
-    if (!IsPlatformMockICD()) {
-        int* data = static_cast<int*>(buffer.Memory().Map());
-        for (uint32_t i = 0; i < 4; ++i) {
-            ASSERT_EQ(data[i], color.int32[i]);
-        }
-    }
-}
-
-// Currently crashing in latest NVIDIA driver (should have a fix, need to test)
-// Confirms works on AMD/Intel Mesa
-TEST_F(PositiveDescriptorHeap, DISABLED_SecondaryCmdBufferCompute) {
-    AddRequiredExtensions(VK_KHR_SHADER_UNTYPED_POINTERS_EXTENSION_NAME);
-    AddRequiredFeature(vkt::Feature::shaderUntypedPointers);
-    RETURN_IF_SKIP(InitBasicDescriptorHeap());
-
-    const VkDeviceSize resource_stride = heap_props.bufferDescriptorSize;
-    CreateResourceHeap(resource_stride);
-
-    vkt::Buffer buffer(*m_device, 256, VK_BUFFER_USAGE_2_STORAGE_BUFFER_BIT_KHR, vkt::device_address);
-
-    VkHostAddressRangeEXT descriptor_host = {resource_heap_data_, static_cast<size_t>(resource_stride)};
-    VkDeviceAddressRangeEXT device_range = buffer.AddressRange();
-
-    VkResourceDescriptorInfoEXT descriptor_info = vku::InitStructHelper();
-    descriptor_info.type = VK_DESCRIPTOR_TYPE_STORAGE_BUFFER;
-    descriptor_info.data.pAddressRange = &device_range;
-
-    vk::WriteResourceDescriptorsEXT(*m_device, 1u, &descriptor_info, &descriptor_host);
-
-    char const* cs_source = R"glsl(
-        #version 450
-        #extension GL_EXT_descriptor_heap : require
-        layout(local_size_x = 1) in;
-        layout(descriptor_heap) buffer A { uint a; } heap[];
-        layout(push_constant) uniform PushConstant {
-            uint b;
-        };
-        void main() {
-            heap[0].a = b;
-        }
-    )glsl";
-    VkShaderObj cs_module = VkShaderObj(*m_device, cs_source, VK_SHADER_STAGE_COMPUTE_BIT, SPV_ENV_VULKAN_1_2);
-
-    VkPipelineCreateFlags2CreateInfoKHR pipeline_create_flags_2_create_info = vku::InitStructHelper();
-    pipeline_create_flags_2_create_info.flags = VK_PIPELINE_CREATE_2_DESCRIPTOR_HEAP_BIT_EXT;
-
-    CreateComputePipelineHelper pipe(*this, &pipeline_create_flags_2_create_info);
-    pipe.cp_ci_.layout = VK_NULL_HANDLE;
-    pipe.cp_ci_.stage = cs_module.GetStageCreateInfo();
-    pipe.CreateComputePipeline(false);
-
-    uint32_t src_data = 4321u;
-
-    VkPushDataInfoEXT push_data_info = vku::InitStructHelper();
-    push_data_info.offset = 0u;
-    push_data_info.data.size = sizeof(uint32_t);
-    push_data_info.data.address = &src_data;
-
-    VkPushConstantRange push_const_range = {VK_SHADER_STAGE_COMPUTE_BIT, 0, sizeof(uint32_t)};
-    VkPipelineLayoutCreateInfo pipeline_layout_info = vku::InitStructHelper();
-    pipeline_layout_info.pushConstantRangeCount = 1u;
-    pipeline_layout_info.pPushConstantRanges = &push_const_range;
-    vkt::PipelineLayout pipeline_layout(*m_device, pipeline_layout_info);
-
-    vkt::CommandBuffer secondary(*m_device, m_command_pool, VK_COMMAND_BUFFER_LEVEL_SECONDARY);
-
-    VkBindHeapInfoEXT bind_resource_info = vku::InitStructHelper();
-    bind_resource_info.heapRange = resource_heap_.AddressRange();
-    bind_resource_info.reservedRangeOffset = resource_heap_.CreateInfo().size - heap_props.minResourceHeapReservedRange;
-    bind_resource_info.reservedRangeSize = heap_props.minResourceHeapReservedRange;
-
-    VkCommandBufferInheritanceDescriptorHeapInfoEXT inheritance_heap_info = vku::InitStructHelper();
-    inheritance_heap_info.pResourceHeapBindInfo = &bind_resource_info;
-
-    VkCommandBufferInheritanceInfo inheritance_info = vku::InitStructHelper(&inheritance_heap_info);
-
-    VkCommandBufferBeginInfo begin_info = vku::InitStructHelper();
-    begin_info.flags = VK_COMMAND_BUFFER_USAGE_ONE_TIME_SUBMIT_BIT;
-    begin_info.pInheritanceInfo = &inheritance_info;
-
-    secondary.Begin(&begin_info);
-    vk::CmdBindPipeline(secondary, VK_PIPELINE_BIND_POINT_COMPUTE, pipe);
-    vk::CmdPushDataEXT(secondary, &push_data_info);
-    vk::CmdDispatch(secondary, 1, 1, 1);
-    secondary.End();
-
-    m_command_buffer.Begin();
-    vk::CmdBindResourceHeapEXT(m_command_buffer, &bind_resource_info);
-    vk::CmdExecuteCommands(m_command_buffer, 1, &secondary.handle());
-    m_command_buffer.End();
-    m_default_queue->SubmitAndWait(m_command_buffer);
-
-    if (!IsPlatformMockICD()) {
-        uint32_t* data = static_cast<uint32_t*>(buffer.Memory().Map());
-        ASSERT_EQ(data[0], src_data);
-    }
-}
-
-TEST_F(PositiveDescriptorHeap, SecondaryCmdBufferGraphics) {
-    AddRequiredExtensions(VK_KHR_SHADER_UNTYPED_POINTERS_EXTENSION_NAME);
-    AddRequiredFeature(vkt::Feature::shaderUntypedPointers);
-    RETURN_IF_SKIP(InitBasicDescriptorHeap());
-    InitRenderTarget();
-
-    const VkDeviceSize resource_stride = heap_props.bufferDescriptorSize;
-    CreateResourceHeap(resource_stride);
-
-    vkt::Buffer buffer(*m_device, 256, VK_BUFFER_USAGE_2_STORAGE_BUFFER_BIT_KHR, vkt::device_address);
-
-    VkHostAddressRangeEXT descriptor_host = {resource_heap_data_, static_cast<size_t>(resource_stride)};
-    VkDeviceAddressRangeEXT device_range = buffer.AddressRange();
-    VkResourceDescriptorInfoEXT descriptor_info = vku::InitStructHelper();
-    descriptor_info.type = VK_DESCRIPTOR_TYPE_STORAGE_BUFFER;
-    descriptor_info.data.pAddressRange = &device_range;
-
-    vk::WriteResourceDescriptorsEXT(*m_device, 1u, &descriptor_info, &descriptor_host);
-
-    char const* vs_source = R"glsl(
-        #version 450
-        #extension GL_EXT_descriptor_heap : require
-
-        layout(descriptor_heap) buffer A { uint a; } heap[];
-        layout(push_constant) uniform PushConstant {
-            uint b;
-        };
-        void main() {
-            heap[0].a = b;
-            gl_Position = vec4(1.0f);
-        }
-    )glsl";
-    VkShaderObj vert_module = VkShaderObj(*m_device, vs_source, VK_SHADER_STAGE_VERTEX_BIT, SPV_ENV_VULKAN_1_2);
-
-    VkPipelineCreateFlags2CreateInfoKHR pipeline_create_flags_2_create_info = vku::InitStructHelper();
-    pipeline_create_flags_2_create_info.flags = VK_PIPELINE_CREATE_2_DESCRIPTOR_HEAP_BIT_EXT;
-
-    VkPipelineShaderStageCreateInfo stage;
-    stage = vert_module.GetStageCreateInfo();
-
-    CreatePipelineHelper pipe(*this, &pipeline_create_flags_2_create_info);
-    pipe.gp_ci_.layout = VK_NULL_HANDLE;
-    pipe.gp_ci_.stageCount = 1;
-    pipe.gp_ci_.pStages = &stage;
-    pipe.CreateGraphicsPipeline(false);
-
-    uint32_t src_data = 4321u;
-
-    VkPushDataInfoEXT push_data_info = vku::InitStructHelper();
-    push_data_info.offset = 0u;
-    push_data_info.data.size = sizeof(uint32_t);
-    push_data_info.data.address = &src_data;
-
-    VkPushConstantRange push_const_range = {VK_SHADER_STAGE_COMPUTE_BIT, 0, sizeof(uint32_t)};
-    VkPipelineLayoutCreateInfo pipeline_layout_info = vku::InitStructHelper();
-    pipeline_layout_info.pushConstantRangeCount = 1u;
-    pipeline_layout_info.pPushConstantRanges = &push_const_range;
-    vkt::PipelineLayout pipeline_layout(*m_device, pipeline_layout_info);
-
-    vkt::CommandBuffer secondary(*m_device, m_command_pool, VK_COMMAND_BUFFER_LEVEL_SECONDARY);
-
-    VkBindHeapInfoEXT bind_resource_info = vku::InitStructHelper();
-    bind_resource_info.heapRange = resource_heap_.AddressRange();
-    bind_resource_info.reservedRangeOffset = resource_heap_.CreateInfo().size - heap_props.minResourceHeapReservedRange;
-    bind_resource_info.reservedRangeSize = heap_props.minResourceHeapReservedRange;
-
-    VkCommandBufferInheritanceDescriptorHeapInfoEXT inheritance_heap_info = vku::InitStructHelper();
-    inheritance_heap_info.pResourceHeapBindInfo = &bind_resource_info;
-
-    VkCommandBufferInheritanceInfo inheritance_info = vku::InitStructHelper(&inheritance_heap_info);
-    inheritance_info.renderPass = RenderPass();
-    inheritance_info.framebuffer = Framebuffer();
-
-    VkCommandBufferBeginInfo begin_info = vku::InitStructHelper();
-    begin_info.flags = VK_COMMAND_BUFFER_USAGE_RENDER_PASS_CONTINUE_BIT;
-    begin_info.pInheritanceInfo = &inheritance_info;
-
-    secondary.Begin(&begin_info);
-    vk::CmdBindPipeline(secondary, VK_PIPELINE_BIND_POINT_GRAPHICS, pipe);
-    vk::CmdPushDataEXT(secondary, &push_data_info);
-    vk::CmdDraw(secondary, 3u, 1u, 0u, 0u);
-    secondary.End();
-
-    m_command_buffer.Begin();
-    vk::CmdBindResourceHeapEXT(m_command_buffer, &bind_resource_info);
-    m_command_buffer.BeginRenderPass(m_renderPassBeginInfo, VK_SUBPASS_CONTENTS_SECONDARY_COMMAND_BUFFERS);
-    vk::CmdExecuteCommands(m_command_buffer, 1, &secondary.handle());
-    m_command_buffer.EndRenderPass();
-    m_command_buffer.End();
-    m_default_queue->SubmitAndWait(m_command_buffer);
-
-    if (!IsPlatformMockICD()) {
-        uint32_t* data = static_cast<uint32_t*>(buffer.Memory().Map());
-        ASSERT_EQ(data[0], src_data);
-    }
-}
-
-TEST_F(PositiveDescriptorHeap, HardcodedOffsetIntoStruct) {
-    AddRequiredExtensions(VK_KHR_SHADER_UNTYPED_POINTERS_EXTENSION_NAME);
-    AddRequiredFeature(vkt::Feature::shaderUntypedPointers);
-    RETURN_IF_SKIP(InitBasicDescriptorHeap());
-
-    const VkDeviceSize resource_stride = heap_props.bufferDescriptorSize;
-    if (resource_stride != 64) {
-        GTEST_SKIP() << "SPIR-V hardcoded for bufferDescriptorSize of 64";
-    }
-    CreateResourceHeap(resource_stride * 4);
-
-    vkt::Buffer buffer_a(*m_device, 256, VK_BUFFER_USAGE_2_STORAGE_BUFFER_BIT_KHR, vkt::device_address);
-    vkt::Buffer buffer_b(*m_device, 256, VK_BUFFER_USAGE_2_STORAGE_BUFFER_BIT_KHR, vkt::device_address);
-    uint8_t host_decriptor_a[kMaxSSBO];
-    uint8_t host_decriptor_b[kMaxSSBO];
-    VkHostAddressRangeEXT descriptor_host = {host_decriptor_a, static_cast<size_t>(resource_stride)};
-    VkDeviceAddressRangeEXT device_range = {buffer_a.Address(), 256};
-    VkResourceDescriptorInfoEXT descriptor_info = vku::InitStructHelper();
-    descriptor_info.type = VK_DESCRIPTOR_TYPE_STORAGE_BUFFER;
-    descriptor_info.data.pAddressRange = &device_range;
-    vk::WriteResourceDescriptorsEXT(*m_device, 1u, &descriptor_info, &descriptor_host);
-
-    device_range.address = buffer_b.Address();
-    descriptor_host.address = host_decriptor_b;
-    vk::WriteResourceDescriptorsEXT(*m_device, 1u, &descriptor_info, &descriptor_host);
-
-    // Set |A| to descriptor index [0, 2, 3] and |B| to [1]
-    memcpy(&resource_heap_data_[0], host_decriptor_a, static_cast<size_t>(resource_stride));
-    memcpy(&resource_heap_data_[resource_stride], host_decriptor_b, static_cast<size_t>(resource_stride));
-    memcpy(&resource_heap_data_[resource_stride * 2], host_decriptor_a, static_cast<size_t>(resource_stride));
-    memcpy(&resource_heap_data_[resource_stride * 3], host_decriptor_a, static_cast<size_t>(resource_stride));
-
-    // What the shader looks like
-    //
-    // layout(descriptor_heap) struct {
-    //    layout(offset = 0)   buffer a { uint data; } x;
-    //    layout(offset = 128) buffer a { uint data; } y;
-    //    layout(offset = 64)  buffer b { uint data; } z;
-    //    layout(offset = 192) buffer a { uint data; } w;
-    // };
-    // layout(push_constant) uniform PushConstant { uint payload; };
-    // void main() {
-    //     heap.z.data = payload;
-    // }
-    char const* cs_source = R"(
-               OpCapability Shader
-               OpCapability UntypedPointersKHR
-               OpCapability DescriptorHeapEXT
-               OpExtension "SPV_EXT_descriptor_heap"
-               OpExtension "SPV_KHR_untyped_pointers"
-               OpMemoryModel Logical GLSL450
-               OpEntryPoint GLCompute %main "main" %resource_heap %_
-               OpExecutionMode %main LocalSize 1 1 1
-               OpDecorate %resource_heap BuiltIn ResourceHeapEXT
-               OpDecorate %A Block
-               OpMemberDecorate %A 0 Offset 0
-               OpDecorate %PushConstant Block
-               OpMemberDecorate %PushConstant 0 Offset 0
-
-               ; Use hardcoded offset
-               OpMemberDecorate %heap_struct 0 Offset 0
-               OpMemberDecorate %heap_struct 1 Offset 128
-               OpMemberDecorate %heap_struct 2 Offset 64
-               OpMemberDecorate %heap_struct 3 Offset 192
-
-       %void = OpTypeVoid
-          %3 = OpTypeFunction %void
-%_ptr_UniformConstant = OpTypeUntypedPointerKHR UniformConstant
-%resource_heap = OpUntypedVariableKHR %_ptr_UniformConstant UniformConstant
-        %int = OpTypeInt 32 1
-      %int_2 = OpConstant %int 2
-       %uint = OpTypeInt 32 0
-          %A = OpTypeStruct %uint
-      %int_0 = OpConstant %int 0
-%PushConstant = OpTypeStruct %uint
-%_ptr_PushConstant_PushConstant = OpTypePointer PushConstant %PushConstant
-          %_ = OpVariable %_ptr_PushConstant_PushConstant PushConstant
-%_ptr_PushConstant_uint = OpTypePointer PushConstant %uint
-%_ptr_StorageBuffer = OpTypeUntypedPointerKHR StorageBuffer
-         %21 = OpTypeBufferEXT StorageBuffer
-%heap_struct = OpTypeStruct %21 %21 %21 %21
-       %main = OpFunction %void None %3
-          %5 = OpLabel
-         %17 = OpAccessChain %_ptr_PushConstant_uint %_ %int_0
-         %18 = OpLoad %uint %17
-         %20 = OpUntypedAccessChainKHR %_ptr_UniformConstant %heap_struct %resource_heap %int_2
-         %24 = OpBufferPointerEXT %_ptr_StorageBuffer %20
-         %25 = OpUntypedAccessChainKHR %_ptr_StorageBuffer %A %24 %int_0
-               OpStore %25 %18
-               OpReturn
-               OpFunctionEnd
-
-    )";
-    VkShaderObj cs_module = VkShaderObj(*m_device, cs_source, VK_SHADER_STAGE_COMPUTE_BIT, SPV_ENV_VULKAN_1_3, SPV_SOURCE_ASM);
-
-    VkPipelineCreateFlags2CreateInfoKHR pipeline_create_flags_2_create_info = vku::InitStructHelper();
-    pipeline_create_flags_2_create_info.flags = VK_PIPELINE_CREATE_2_DESCRIPTOR_HEAP_BIT_EXT;
-
-    CreateComputePipelineHelper pipe(*this, &pipeline_create_flags_2_create_info);
-    pipe.cp_ci_.layout = VK_NULL_HANDLE;
-    pipe.cp_ci_.stage = cs_module.GetStageCreateInfo();
-    pipe.CreateComputePipeline(false);
-
-    uint32_t src_data = 4321u;
-
-    VkPushDataInfoEXT push_data_info = vku::InitStructHelper();
-    push_data_info.offset = 0u;
-    push_data_info.data.size = sizeof(uint32_t);
-    push_data_info.data.address = &src_data;
-
-    VkPushConstantRange push_const_range = {VK_SHADER_STAGE_COMPUTE_BIT, 0, sizeof(uint32_t)};
-    VkPipelineLayoutCreateInfo pipeline_layout_info = vku::InitStructHelper();
-    pipeline_layout_info.pushConstantRangeCount = 1u;
-    pipeline_layout_info.pPushConstantRanges = &push_const_range;
-    vkt::PipelineLayout pipeline_layout(*m_device, pipeline_layout_info);
-
-    m_command_buffer.Begin();
-    vk::CmdPushConstants(m_command_buffer, pipeline_layout.handle(), VK_SHADER_STAGE_COMPUTE_BIT, 0, sizeof(uint32_t), &src_data);
-    BindResourceHeap();
-    vk::CmdPushDataEXT(m_command_buffer, &push_data_info);
-    vk::CmdBindPipeline(m_command_buffer, VK_PIPELINE_BIND_POINT_COMPUTE, pipe);
-    vk::CmdDispatch(m_command_buffer, 1, 1, 1);
-    m_command_buffer.End();
-    m_default_queue->SubmitAndWait(m_command_buffer);
-
-    if (!IsPlatformMockICD()) {
-        uint32_t* data = static_cast<uint32_t*>(buffer_b.Memory().Map());
-        ASSERT_EQ(data[0], src_data);
-    }
-}
-
-TEST_F(PositiveDescriptorHeap, SingleElementNoArray) {
-    TEST_DESCRIPTION("GLSL only allows arrays of descriptor, force it to be a single element");
-    AddRequiredExtensions(VK_KHR_SHADER_UNTYPED_POINTERS_EXTENSION_NAME);
-    AddRequiredFeature(vkt::Feature::shaderUntypedPointers);
-    RETURN_IF_SKIP(InitBasicDescriptorHeap());
-
-    const VkDeviceSize resource_stride = heap_props.bufferDescriptorSize;
-    CreateResourceHeap(resource_stride);
-
-    vkt::Buffer buffer(*m_device, 256, VK_BUFFER_USAGE_2_STORAGE_BUFFER_BIT_KHR, vkt::device_address);
-    VkHostAddressRangeEXT descriptor_host = {resource_heap_data_, static_cast<size_t>(resource_stride)};
-    VkDeviceAddressRangeEXT device_range = buffer.AddressRange();
-    VkResourceDescriptorInfoEXT descriptor_info = vku::InitStructHelper();
-    descriptor_info.type = VK_DESCRIPTOR_TYPE_STORAGE_BUFFER;
-    descriptor_info.data.pAddressRange = &device_range;
-    vk::WriteResourceDescriptorsEXT(*m_device, 1u, &descriptor_info, &descriptor_host);
-
-    // GLSL can go
-    //    layout(descriptor_heap) buffer A { uint a; } heap[];
-    // but this shader looks like
-    //
-    //     layout(descriptor_heap) buffer A { uint a; } heap;
-    //     layout(push_constant) uniform PushConstant { uint b; };
-    //     void main() {
-    //         heap.a = b;
-    //     }
-    char const* cs_source = R"(
-               OpCapability Shader
-               OpCapability UntypedPointersKHR
-               OpCapability DescriptorHeapEXT
-               OpExtension "SPV_EXT_descriptor_heap"
-               OpExtension "SPV_KHR_untyped_pointers"
-               OpMemoryModel Logical GLSL450
-               OpEntryPoint GLCompute %main "main" %resource_heap %_
-               OpExecutionMode %main LocalSize 1 1 1
-               OpDecorate %resource_heap BuiltIn ResourceHeapEXT
-               OpDecorate %A Block
-               OpMemberDecorate %A 0 Offset 0
-               OpDecorate %PushConstant Block
-               OpMemberDecorate %PushConstant 0 Offset 0
-       %void = OpTypeVoid
-          %3 = OpTypeFunction %void
-%_ptr_UniformConstant = OpTypeUntypedPointerKHR UniformConstant
-%resource_heap = OpUntypedVariableKHR %_ptr_UniformConstant UniformConstant
-        %int = OpTypeInt 32 1
-      %int_0 = OpConstant %int 0
-       %uint = OpTypeInt 32 0
-          %A = OpTypeStruct %uint
-%PushConstant = OpTypeStruct %uint
-%_ptr_PushConstant_PushConstant = OpTypePointer PushConstant %PushConstant
-          %_ = OpVariable %_ptr_PushConstant_PushConstant PushConstant
-%_ptr_PushConstant_uint = OpTypePointer PushConstant %uint
-%_ptr_StorageBuffer = OpTypeUntypedPointerKHR StorageBuffer
-         %20 = OpTypeBufferEXT StorageBuffer
-       %main = OpFunction %void None %3
-          %5 = OpLabel
-         %16 = OpAccessChain %_ptr_PushConstant_uint %_ %int_0
-         %17 = OpLoad %uint %16
-         %19 = OpUntypedAccessChainKHR %_ptr_UniformConstant %20 %resource_heap
-         %23 = OpBufferPointerEXT %_ptr_StorageBuffer %19
-         %24 = OpUntypedAccessChainKHR %_ptr_StorageBuffer %A %23 %int_0
-               OpStore %24 %17
-               OpReturn
-               OpFunctionEnd
-    )";
-    VkShaderObj cs_module = VkShaderObj(*m_device, cs_source, VK_SHADER_STAGE_COMPUTE_BIT, SPV_ENV_VULKAN_1_3, SPV_SOURCE_ASM);
-
-    VkPipelineCreateFlags2CreateInfoKHR pipeline_create_flags_2_create_info = vku::InitStructHelper();
-    pipeline_create_flags_2_create_info.flags = VK_PIPELINE_CREATE_2_DESCRIPTOR_HEAP_BIT_EXT;
-
-    CreateComputePipelineHelper pipe(*this, &pipeline_create_flags_2_create_info);
-    pipe.cp_ci_.layout = VK_NULL_HANDLE;
-    pipe.cp_ci_.stage = cs_module.GetStageCreateInfo();
-    pipe.CreateComputePipeline(false);
-
-    uint32_t src_data = 4321u;
-
-    VkPushDataInfoEXT push_data_info = vku::InitStructHelper();
-    push_data_info.offset = 0u;
-    push_data_info.data.size = sizeof(uint32_t);
-    push_data_info.data.address = &src_data;
-
-    VkPushConstantRange push_const_range = {VK_SHADER_STAGE_COMPUTE_BIT, 0, sizeof(uint32_t)};
-    VkPipelineLayoutCreateInfo pipeline_layout_info = vku::InitStructHelper();
-    pipeline_layout_info.pushConstantRangeCount = 1u;
-    pipeline_layout_info.pPushConstantRanges = &push_const_range;
-    vkt::PipelineLayout pipeline_layout(*m_device, pipeline_layout_info);
-
-    m_command_buffer.Begin();
-    vk::CmdPushConstants(m_command_buffer, pipeline_layout.handle(), VK_SHADER_STAGE_COMPUTE_BIT, 0, sizeof(uint32_t), &src_data);
-    BindResourceHeap();
-    vk::CmdPushDataEXT(m_command_buffer, &push_data_info);
-    vk::CmdBindPipeline(m_command_buffer, VK_PIPELINE_BIND_POINT_COMPUTE, pipe);
-    vk::CmdDispatch(m_command_buffer, 1, 1, 1);
-    m_command_buffer.End();
-    m_default_queue->SubmitAndWait(m_command_buffer);
-
-    if (!IsPlatformMockICD()) {
-        uint32_t* data = static_cast<uint32_t*>(buffer.Memory().Map());
-        ASSERT_EQ(data[0], src_data);
     }
 }
